requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# COREUS Framework dependencies
langgraph>=0.0.20

# Web scraping
Crawl4AI>=0.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Email and templating
jinja2>=3.1.2
email-validator>=2.0.0
aiosmtplib>=3.0.0

# LLM
openai>=1.12.0

# Scheduling
apscheduler>=3.10.4

# Database
google-cloud-firestore>=2.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

